Variables

+++ runtime/print_int.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int print_int(long x) {
-    return printf("%ld\n", x);
+int print_int(long long x) {
+    return printf("%lld\n", x);
 }
